Delete User entity and update Client entity
