--- src/DataFixtures/UserFixtures.php
@@ -1,26 +1,0 @@
-<?php
-
-namespace App\DataFixtures;
-
-use Faker\Factory;
-use App\Entity\User;
-use Doctrine\Persistence\ObjectManager;
-use Doctrine\Bundle\FixturesBundle\Fixture;
-use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
-
-class UserFixtures extends Fixture
-{
-    public function load(ObjectManager $manager)
-    {
-        $faker = \Faker\Factory::create('fr_FR');
-
-        $admin = new User();
-        $admin->setEmail('admin@bilemo.com')
-            ->setCreatedAt($faker->dateTime())
-            ->setPassword($faker->password)
-            ->setRoles(['ROLE_ADMIN']);
-
-        $manager->persist($admin);
-        $manager->flush();
-    }
-}
